Add association in parameter
